requirements.txt
anthropic>=0.25.0
supabase>=2.3.0
lark-oapi>=1.3.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
yt-dlp>=2024.1.0
rapidfuzz>=3.6.0
python-dotenv>=1.0.0
requests-cache>=1.2.0
tenacity>=8.2.0
